hospital_node/requirements.txt
# Federated Learning
flwr==1.6.0

# Machine Learning
tensorflow==2.15.0
numpy==1.24.3
scikit-learn==1.3.2
Pillow==10.2.0
opencv-python==4.9.0.80

# Communication
grpcio==1.48.2
grpcio-tools==1.48.2
protobuf==3.20.3

# Encryption
phe==1.5.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Logging
loguru==0.7.2
